pyproject.toml
[project]
name = "MAS"
version = "0.0.1"
description = "Starter template for making a custom Reasoning and Action agent (using tool calling) in LangGraph."
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.3.27",
    "langgraph-sdk>=0.1.66",
    "langgraph-checkpoint>=2.0.23",
    "langchain-core>=0.3,<0.4",
    "langsmith>=0.3.45,<0.4",
    "openai",
    "anthropic",
    "together",
    "numpy",
    "tqdm",
    "backoff",
    "python-dotenv",
    "pandas",
    "jinja2",
    "pydantic",
    "aiosqlite",
    "langchain>=0.2.14",
    "langchain-openai>=0.1.22",
    "langchain-anthropic>=0.1.23",
    "langchain-fireworks>=0.1.7",
    "langchain-community",
    "pymilvus[model]>=2.3.7",
    "langchain-milvus",
    "google-genai",
    "unstructured",
    "beautifulsoup4",
    "requests",
    "pillow",
    "pypdf",
    "python-docx",
    "python-pptx",
    "openpyxl",
    "PyMuPDF",
    "tabula-py",
    "tavily-python",
    "dashscope",
    "moviepy==1.0.3",
    "opencv-python>=4.8.0",
    "imageio>=2.31.0",
    "pydub>=0.25.1",
    "mutagen>=1.46.0",
    "psutil>=5.9.0",
    "pytz>=2025.2",
    "uvicorn>=0.35.0",
    "fastapi>=0.116.1",
    "fastapi-events>=0.12.2",
    "websockets>=15.0.1",
    # LangGraph CLI 相关依赖 - 解决版本冲突
    "langgraph-cli>=0.3.6",
    "langgraph-api>=0.2.130",
    "langgraph-runtime-inmem>=0.6.13",
    # 降级 sse-starlette 以兼容 langgraph-api
    "sse-starlette>=2.1.0,<2.2.0",
]

[project.optional-dependencies]
# 开发环境依赖
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

# 本地开发服务器依赖
local = [
    "langgraph-cli[inmem]>=0.3.6",
]

[tool.setuptools.packages.find]
include = ["*"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# 添加 uv 配置
[tool.uv]
# 设置依赖解析策略
resolution = "lowest"
